# storage: add conflict handling for MVCC range tombstones