This is for the `mvcc-range-tombstones` branch. Only the last commit should be reviewed here, the others are from #78946 and its ancestors.

---

This patch makes sure all MVCC write operations take range tombstones
into account for conflict checks and conditions.

Release note: None