@erikgrinaker 2022-05-30
@jbowens I think we should focus our optimization efforts here first. This patch enables conflict checks with MVCC range tombstones on the write path. For `MVCCPut` on a database with no range keys or other existing data, this has a ~6% performance penalty:

```
name                               old time/op    new time/op    delta
MVCCPut_Pebble/valueSize=10-24       9.82µs ± 6%   10.42µs ± 5%  +6.05%  (p=0.001 n=10+10)
MVCCPut_Pebble/valueSize=100-24      11.0µs ± 6%    11.3µs ± 2%  +2.97%  (p=0.013 n=9+10)
MVCCPut_Pebble/valueSize=1000-24     12.7µs ± 2%    13.4µs ± 3%  +5.63%  (p=0.000 n=10+9)
MVCCPut_Pebble/valueSize=10000-24    42.7µs ±15%    40.5µs ±13%    ~     (p=0.143 n=10+10)
```

The conflict checks essentially enable `IterKeyTypePointsAndRanges` and then use `pointSynthesizingIter` to synthesize point tombstones. The latter is not optimized, so there are some gains to be had there, but the majority appears to be due to Pebble -- disabling point synthesis (so we're only changing `IterKeyType`) still shows a ~5% performance penalty:

```
name                            old time/op    new time/op    delta
MVCCPut_Pebble/valueSize=10-24    9.82µs ± 6%   10.28µs ± 6%  +4.69%  (p=0.009 n=10+10)
```

Ideally, we should be able to get this down to ~0 in the null case (no range keys). That would allow us to merge this to `master` and unblock dependant PRs.