@erikgrinaker 2022-05-30
I think most of the Pebble overhead here may be due to iterator construction. When I change the benchmark to use a batch, which reuses the Pebble iterator, then the time/op drops by an order of magnitude. In that case, the overhead is mostly due to `pointSynthesizingIter`.

```
name                            old time/op    new time/op   delta
MVCCPut_Pebble/valueSize=10-24     910ns ± 4%   1034ns ± 2%  +13.61%  (p=0.008 n=5+5)
```

When I drop the point synthesis and only enable range keys on the iterator, the performance diff is negligible:

```
name                            old time/op    new time/op    delta
MVCCPut_Pebble/valueSize=10-24     910ns ± 4%     933ns ± 5%   ~     (p=0.095 n=5+5)
```

I've pushed a patch with some WIP `pointSynthesizingIter` optimizations here, I'll see how low we can get it. It's possible that we're better off explicitly handling range keys in `mvccPutInternal`.

In any case, it would be worth having a look at cutting the iterator construction costs too, although this may not be as important since we often reuse iterators on batches.