@erikgrinaker 2022-06-01
I eliminated `pointSynthesisIter` on the hottest write paths, only using it when we need to access an old value. This got rid of most of the performance diff. We now only have a minor regression when updating existing keys, and only when using a batch (which reuses the iterator).

I think this is sufficient to merge to `master`, and we can keep pushing on performance separately.

```
name                                                       old time/op    new time/op    delta
MVCCPut_Pebble/batch=false,valueSize=10,versions=1-24        10.1µs ± 7%     9.9µs ± 5%    ~     (p=0.211 n=10+9)
MVCCPut_Pebble/batch=false,valueSize=10,versions=10-24        113µs ± 9%     111µs ± 3%    ~     (p=0.529 n=10+10)
MVCCPut_Pebble/batch=false,valueSize=100,versions=1-24       11.2µs ± 5%    11.0µs ± 5%    ~     (p=0.182 n=9+10)
MVCCPut_Pebble/batch=false,valueSize=100,versions=10-24       124µs ± 4%     124µs ± 3%    ~     (p=0.853 n=10+10)
MVCCPut_Pebble/batch=false,valueSize=1000,versions=1-24      12.7µs ± 5%    12.7µs ± 4%    ~     (p=0.955 n=10+10)
MVCCPut_Pebble/batch=false,valueSize=1000,versions=10-24      155µs ± 2%     150µs ± 3%  -3.22%  (p=0.000 n=9+10)
MVCCPut_Pebble/batch=false,valueSize=10000,versions=1-24     41.2µs ± 6%    41.7µs ± 8%    ~     (p=0.529 n=10+10)
MVCCPut_Pebble/batch=false,valueSize=10000,versions=10-24     449µs ± 6%     451µs ± 9%    ~     (p=0.684 n=10+10)
MVCCPut_Pebble/batch=true,valueSize=10,versions=1-24          952ns ± 7%     969ns ± 4%    ~     (p=0.696 n=10+8)
MVCCPut_Pebble/batch=true,valueSize=10,versions=10-24        15.2µs ± 1%    15.6µs ± 1%  +2.60%  (p=0.000 n=9+10)
MVCCPut_Pebble/batch=true,valueSize=100,versions=1-24        1.09µs ± 2%    1.08µs ± 1%    ~     (p=0.251 n=9+10)
MVCCPut_Pebble/batch=true,valueSize=100,versions=10-24       16.7µs ± 2%    17.1µs ± 3%  +2.81%  (p=0.000 n=10+10)
MVCCPut_Pebble/batch=true,valueSize=1000,versions=1-24       1.92µs ± 2%    1.90µs ± 2%    ~     (p=0.130 n=9+9)
MVCCPut_Pebble/batch=true,valueSize=1000,versions=10-24      29.2µs ± 2%    29.7µs ± 1%  +1.91%  (p=0.001 n=9+10)
MVCCPut_Pebble/batch=true,valueSize=10000,versions=1-24      10.7µs ±14%    10.2µs ±12%    ~     (p=0.190 n=10+10)
MVCCPut_Pebble/batch=true,valueSize=10000,versions=10-24      122µs ± 7%     123µs ± 7%    ~     (p=0.912 n=10+10)
```