@jbowens 2022-06-03
Sorry for the delay getting back to you—had a run around with covid this week. A bit of it might be fixed by cockroachdb/pebble#1723. I suspect the rest _is_ in iterator construction, and I'm pretty confident it'll be easy to fix. I'll work on it right now.