@erikgrinaker 2022-06-04
> Sorry for the delay getting back to you—had a run around with covid this week. A bit of it might be fixed by [cockroachdb/pebble#1723](https://github.com/cockroachdb/pebble/pull/1723). I suspect the rest _is_ in iterator construction, and I'm pretty confident it'll be easy to fix. I'll work on it right now.

Oh no, hope the covid wasn't too bad.

I think we're good for this PR. The read path is a bit worse, let's continue over on #82045.